fixed defaultpasswordnumber, should be 2^30 to achieve max efficiency

+++ src/gpu/hash/commons.cuh
@@ -18,7 +18,7 @@ __host__ double memoryAnalysis(int passwordNumber);
 __host__ int computeBatchSize(double numberOfPass, int passwordNumber);
 __host__ void kernel(double numberOfPass, int batchSize, float * milliseconds, const clock_t * program_start,
                      Digest ** h_results, Password **h_passwords, int passwordNumber);
-__host__ Password * generatePasswords(long passwordNumber);
+__host__ Password * generatePasswords(int passwordNumber);
 
 
 #endif //CUDA_NAIVE_EXHAUSTIVE_SEARCH_COMMONS_CUH
